Added handling of all base stats, fixed some issues

+++ Assets/Scripts/Scriptable Objects/Perks/HealthIncreaseEffect.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "HealthIncreaseEffect", menuName = "Perks/Health Increase")]
 public class HealthIncreaseEffect : PerkEffect
 {
-    public int bonusValue = 5;
+    public float bonusValue = 5;
     public override void Apply(PlayerStatsData stats)
     {
         if (stats != null)
